% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_urlkey.R
\name{set_url}
\alias{set_url}
\title{Set url key
Essential function to operate stock screen functions}
\usage{
set_url(filename)
}
\arguments{
\item{filename}{a character vector of the json file}
}
\description{
\code{set_urlkey} sets the url_key for stock screen functions
}
\details{
The function takes a json filename and sets the url_key to the environent
}
